pyproject.toml
[project]
name = "guidecrypt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.2",
    "pycryptodome>=3.23.0",
    "pyinstaller>=6.15.0",
    "tqdm>=4.67.1",
]
